[profile] re-organized code to separate public and private things
